refactor: split search helpers by responsibility
